initial driver item cards created
