--- src/components/side_bar/driver_list/DriverItem2.module.css
@@ -1,107 +1,0 @@
-.accordionsection {
-    display: flex;
-    position: relative;
-    background: white;
-}
-
-.container {
-    border-radius: 6px;
-    position: absolute;
-    top: 30%;
-    width: 100%;
-
-}
-
-.wrap {
-    border-radius: 6px;
-    background: white;
-    display: flex;
-    justify-content: space-between;
-    margin: 0px 1px 1px 1px;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.2);
-}
-
-
-.wrap:hover {
-    background: #fafafa;
-}
-
-.logo {
-    position: relative;
-    top: 10%;
-    padding-left: 5px;
-}
-
-.arrow {
-    padding-top: 7%;
-}
-
-.eta {
-    position: relative;
-    bottom: 15%;
-    padding-top: 2px;
-    padding-left: 5px;
-    color: #0F4C85;
-    font-size: 14px;
-    font-family: sans-serif "open sans";
-}
-
-.cost {
-    position: relative;
-    left: 20%;
-    color: #00305B;
-    font-size: 24px;
-    font-family: sans-serif "open sans";
-
-}
-
-.dropdown {
-    display: inline;
-    background: white;
-    padding-bottom: 10px;
-}
-
-.descriptiontitlediv {
-    margin-left: 2px;
-}
-
-.descriptiontitle {
-    color: #00305B;
-    font-size: 16px;
-    font-family: sans-serif "open sans";
-    margin: 5px 0px 5px 2px;
-}
-
-.divdescription {
-    width: 100%;
-    display: grid;
-    grid-template-columns: 50% 50%;
-}
-
-.title {
-    color: #00305B;
-    font-size: 14px;
-    font-family: sans-serif "open sans";
-    line-height: 20px;
-    margin: 0px 0px 0px 10px;
-}
-
-.ridedata {
-    color: #00305B;
-    font-size: 14px;
-    font-family: sans-serif "open sans";
-    line-height: 20px;
-    margin: 0px 0px 0px 70%;
-    text-align: left;
-    width: 50%;
-}
-
-.ridedatafinal {
-    color: #00305B;
-    font-size: 14px;
-    font-family: sans-serif "open sans";
-    line-height: 20px;
-    margin: 0px 0px 10px 70%;
-    text-align: left;
-    width: 50%;
-}
