+++ src/components/side_bar/driver_list/DriverList.js
@@ -1,15 +1,12 @@
 import React from 'react';
 import DriverItem from './DriverItem';
-import Accordion from './DriverItem2';
 
 import classes from './DriverList.module.css';
 
 function DriverList() {
     return (
         <div className={classes.container}>
-            <ul className={classes.list}>
-                <Accordion></Accordion>
-            </ul>
+            <DriverItem></DriverItem>
         </div>
     );
 }
