+++ src/components/side_bar/driver_list/DriverList.module.css
@@ -4,9 +4,3 @@
     /* box-shadow: 0 2px 8px rgba(0, 0, 0, 0.2); */
     margin: 0px 5px 5px 5px;
 }
-
-.list {
-    list-style: none;
-    margin: 0;
-    padding: 0;
-}
